third_party/blink/web_tests/fast/events/touch/touch-slider-othogonal-movement.html

<!DOCTYPE html>
<title>Orthogonal touch movement should not change slider value</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<input id="slider" type="range" value="0" style="width:200px; height: 50px">
<script>
const w = slider.offsetWidth;
const h = slider.offsetHeight;

test(() => {
  assert_own_property(window, 'eventSender', 'This test needs eventSender to emulate touch');

  // Emulate a touch that starts inside slider and moves vertically.
  const x = slider.offsetLeft + w / 2;
  const startY = slider.offsetTop + h / 2;
  const endY = startY + h;

  eventSender.clearTouchPoints();
  eventSender.addTouchPoint(x, startY);
  eventSender.touchStart();

  eventSender.updateTouchPoint(0, x, endY);
  eventSender.touchMove();

  eventSender.releaseTouchPoint(0);
  eventSender.touchEnd();

  assert_equals(slider.value, "0");
}, 'Orthogonal touch movement should not change slider value');

test(() => {
  assert_own_property(window, 'eventSender', 'This test needs eventSender to emulate touch');

  // Emulate a tap on the slider.
  const x = slider.offsetLeft + w / 2;
  const y = slider.offsetTop + h / 2;

  eventSender.clearTouchPoints();
  eventSender.addTouchPoint(x, y);
  eventSender.touchStart();
  eventSender.releaseTouchPoint(0);
  eventSender.touchEnd();

  assert_not_equals(slider.value, "0");
}, 'Tapping should still change slider value');
</script>
